Moved content of tests/clean/conftest.py to tests/conftest.py
